requirements_mac.txt
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.14.2
blinker==1.4
bs4==0.0.2
cachetools==6.2.0
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
click==8.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==46.0.2
cssselect2==0.8.0
diff-match-patch==20241021
distro==1.7.0
Django==4.2.23
django-apscheduler==0.7.0
django-auditlog==3.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
django-haystack==3.2.1
django-import-export==4.3.10
django-jsonfield==1.4.1
django-mathfilters==1.0.0
django-microsoft-auth==3.0.1
django-model-utils==5.0.0
django-simple-history==3.10.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
future==1.0.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.0.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==8.10.0
numpy==2.2.6
openpyxl==3.1.5
oscrypto==1.3.0
packaging==21.3
pandas==2.3.3
pdfkit==1.0.0
pexpect==4.8.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.23
PyHamcrest==2.0.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.3.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==6.1.1
pypng==0.20220715.0
pyrsistent==0.18.1
pyserial==3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-magic==0.4.24
pytz==2025.2
PyYAML==6.0.3
pyzk==0.9
qrcode==8.2
reportlab==4.0.9
requests==2.32.5
responses==0.25.8
rsa==4.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.17.0
sos==4.8.2
soupsieve==2.8
sqlparse==0.5.3
svglib==1.5.1
swapper==1.4.0
systemd-python==234
tablib==3.8.0
tinycss2==1.4.0
Twisted==22.1.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==1.26.5
utils==1.0.2
wadllib==1.3.6
webencodings==0.5.1
whitenoise==6.11.0
Whoosh==2.7.4
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
zipp==1.0.0
zope.interface==5.4.0
